+++ Sources/SmileID/Classes/DocumentVerification/Model/OrchestratedDocumentVerificationViewModel.swift
@@ -354,10 +354,11 @@ extension IOrchestratedDocumentVerificationViewModel: SmartSelfieResultDelegate 
 // swiftlint:disable opening_brace
 class OrchestratedDocumentVerificationViewModel: IOrchestratedDocumentVerificationViewModel<DocumentVerificationResultDelegate, DocumentVerificationJobResult> {
     override func onFinished(delegate: DocumentVerificationResultDelegate) {
-        if let savedFiles,
+        if let error {
+            delegate.didError(error: error)
+        } else if let savedFiles,
            let selfiePath = getRelativePath(from: selfieFile),
-           let documentFrontPath = getRelativePath(from: savedFiles.documentFront)
-        {
+           let documentFrontPath = getRelativePath(from: savedFiles.documentFront) {
             let documentBackPath = getRelativePath(from: savedFiles.documentBack)
             delegate.didSucceed(
                 selfie: selfiePath,
@@ -365,10 +366,6 @@ class OrchestratedDocumentVerificationViewModel: IOrchestratedDocumentVerificati
                 documentBackImage: documentBackPath,
                 didSubmitDocumentVerificationJob: didSubmitJob
             )
-        } else if let error {
-            // We check error as the 2nd case because as long as jobStatusResponse is not nil, it
-            // was a success
-            delegate.didError(error: error)
         } else {
             delegate.didError(error: SmileIDError.unknown("Unknown error"))
         }
@@ -378,10 +375,11 @@ class OrchestratedDocumentVerificationViewModel: IOrchestratedDocumentVerificati
 // swiftlint:disable opening_brace
 class OrchestratedEnhancedDocumentVerificationViewModel: IOrchestratedDocumentVerificationViewModel<EnhancedDocumentVerificationResultDelegate, EnhancedDocumentVerificationJobResult> {
     override func onFinished(delegate: EnhancedDocumentVerificationResultDelegate) {
-        if let savedFiles,
+        if let error {
+            delegate.didError(error: error)
+        } else if let savedFiles,
            let selfiePath = getRelativePath(from: selfieFile),
-           let documentFrontPath = getRelativePath(from: savedFiles.documentFront)
-        {
+           let documentFrontPath = getRelativePath(from: savedFiles.documentFront) {
             let documentBackPath = getRelativePath(from: savedFiles.documentBack)
             delegate.didSucceed(
                 selfie: selfiePath,
@@ -389,10 +387,6 @@ class OrchestratedEnhancedDocumentVerificationViewModel: IOrchestratedDocumentVe
                 documentBackImage: documentBackPath,
                 didSubmitEnhancedDocVJob: didSubmitJob
             )
-        } else if let error {
-            // We check error as the 2nd case because as long as jobStatusResponse is not nil, it
-            // was a success
-            delegate.didError(error: error)
         } else {
             delegate.didError(error: SmileIDError.unknown("Unknown error"))
         }